+++ src/ts/rendering.ts
@@ -30,13 +30,14 @@ export const orthographicCamera = new THREE.OrthographicCamera(-window.innerWidt
 orthographicCamera.up.set(0, 0, -1);
 orthographicCamera.lookAt(new THREE.Vector3(1, 0, 0));
 
+const dim = 2**9;
 // We use a single texture for the marble reflection and not 6 in a cube because that's 6x faster.
-export const marbleReflectionRenderTarget = new THREE.WebGLRenderTarget(2**12, 2**12, {
+export const marbleReflectionRenderTarget = new THREE.WebGLRenderTarget(dim, dim, {
 	format: THREE.RGBFormat,
 	generateMipmaps: true,
 	minFilter: THREE.LinearMipmapLinearFilter,
 	anisotropy: 16
 });
-export const marbleReflectionCamera = new THREE.PerspectiveCamera(177, 1, 0.05, 1000);
+export const marbleReflectionCamera = new THREE.PerspectiveCamera(150, 1, 0.05, 1000); // Far will be updated on le fly
 
 resize();