+++ version_history.md
@@ -4,6 +4,7 @@
 - Fixed old replays again
 - Fixed incorrect checkpoint orientation in Ultra levels
 - Fixed "Qualify Time"
+- Fixed Tim Trance not playing in all MBU levels
 - Removed Herobrine
 
 ## 2.1.3